requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
base58==2.1.1
bitarray==2.6.2
black==22.12.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
coverage==6.5.0
cytoolz==0.12.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-retry==0.1.15
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
exceptiongroup==1.0.4
fire==0.4.0
frozenlist==1.3.3
hexbytes==0.3.0
idna==3.4
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
jsonschema==4.17.3
lru-dict==1.1.8
multiaddr==0.0.9
multicall==0.7.0
multidict==6.0.4
mypy==0.991
mypy-extensions==0.4.3
netaddr==0.8.0
packaging==22.0
parsimonious==0.8.1
pathspec==0.10.2
platformdirs==2.5.4
pluggy==1.0.0
protobuf==3.19.5
pycryptodome==3.16.0
pydantic==1.10.2
pyrsistent==0.19.3
pytest==7.2.0
pytest-cov==4.0.0
pytest-cover==3.0.0
pytest-coverage==0.0
python-dotenv==0.21.0
requests==2.28.1
rlp==2.0.1
six==1.16.0
termcolor==2.1.1
tinydb==4.7.0
tomli==2.0.1
toolz==0.12.0
types-requests==2.28.11.5
types-urllib3==1.26.25.4
typing_extensions==4.4.0
urllib3==1.26.13
varint==1.0.2
web3==5.31.3
websockets==9.1
yarl==1.8.2
